Final corrections, add types to new files
