+++ src/components/NewMovie/NewMovie.types.ts
@@ -1,0 +1,11 @@
+export interface NewMovieProps {
+  onAdd: (newMovie: Movie) => void;
+}
+
+interface Movie {
+  title: string,
+  description: string,
+  imgUrl: string,
+  imdbUrl: string,
+  imdbId: string,
+}
