+++ src/components/TextField/TextField.types.ts
@@ -1,0 +1,8 @@
+export interface TextFieldProps {
+  name: string;
+  value: string;
+  label?: string;
+  placeholder?: string;
+  required?: boolean;
+  onChange?: (newValue: string) => void;
+}
